api/requirements.txt
fastapi
torch
diffusers
transformers
pillow
uvicorn
python-dotenv
